requirements.txt
# Main -> Requests / Database / Flask
connexion==2.7.*
python-dateutil==2.8.*
setuptools==56.2.*
Flask==1.1.*
waitress==2.0.*
Flask-SQLAlchemy==2.5.*
psycopg2==2.9.*
urllib3==1.26.*

# Metrics + logs
prometheus-flask-exporter==0.18.*
SQLAlchemy==1.4.*
opencensus==0.7.*
opencensus-ext-ocagent==0.7.*
opencensus-ext-flask==0.7.*
opencensus-ext-sqlalchemy==0.1.*
opencensus-ext-requests==0.7.*
opencensus-ext-logging==0.1.*
python-json-logger==2.0.*
coloredlogs==15.0.*

# Kafka
kafka-python==2.0.*

# Openapi and validations
marshmallow==3.13.*
openapi-spec-validator==0.4.*
jsonschema==3.2.*

# Influx
influxdb-client==1.26.0

# Others
PyJWT==2.3.*
pandas==2.0.*
numpy==1.24.*
PyYAML==6.0.1
markupsafe==2.0.1


# hems_auth custom library
# To get the env vars check gitlab: Settings -> CI/CD -> Variables
# Check the HEMS group to retrieve the values
git+https://${GITLAB_DEPLOY_USERNAME}:${GITLAB_DEPLOY_TOKEN}@gitlab.inesctec.pt/cpes/european-projects/interconnect/hems/custom-hems-libraries/hems-auth.git@v2.0.5

# SSA management library
#git+https://${GITLAB_DEPLOY_USERNAME}:${GITLAB_SSA_MANAGER_DEPLOY_TOKEN}@gitlab.inesctec.pt/cpes/european-projects/interconnect/hems/custom-hems-libraries/ssa-manager.git@3.0.1
